// Neptune.Web/src/app/shared/generated/model/field-visit-dto.ts
export class FieldVisitDto { 
    FieldVisitID?: number;
    VisitDate?: string;
    TreatmentBMPID?: number;
    TreatmentBMPName?: string | null;
    TreatmentBMPTypeID?: number;
    TreatmentBMPTypeName?: string | null;
    StormwaterJurisdictionID?: number;
    OrganizationName?: string | null;
    FieldVisitTypeID?: number;
    FieldVisitTypeDisplayName?: string | null;
    PerformedByPersonID?: number;
    PerformedByPersonName?: string | null;
    FieldVisitStatusID?: number;
    FieldVisitStatusDisplayName?: string | null;
    IsFieldVisitVerified?: boolean;
    InventoryUpdated?: boolean;
    NumberOfRequiredAttributes?: number;
    NumberRequiredAttributesEntered?: number;
    readonly RequiredAttributesEntered?: boolean;
    TreatmentBMPAssessmentIDInitial?: number | null;
    IsAssessmentCompleteInitial?: boolean;
    AssessmentScoreInitial?: number | null;
    readonly InitialAssessmentStatus?: string | null;
    TreatmentBMPAssessmentIDPM?: number | null;
    IsAssessmentCompletePM?: boolean;
    AssessmentScorePM?: number | null;
    MaintenanceRecordID?: number | null;
    readonly MaintenanceOccurred?: string | null;
    readonly PostMaintenanceAssessmentStatus?: string | null;
    WaterQualityManagementPlanID?: number | null;
    WaterQualityManagementPlanName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}