// Neptune.Web/src/app/shared/generated/model/funding-event-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FundingEventFundingSourceSimpleDto } from './funding-event-funding-source-simple-dto';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FundingEventDto { 
    FundingEventID?: number;
    TreatmentBMPID?: number;
    FundingEventTypeID?: number;
    Year?: number;
    Description?: string | null;
    FundingEventFundingSources?: Array<FundingEventFundingSourceSimpleDto> | null;
    DisplayName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FundingEventDtoForm { 
    FundingEventID?: FormControl<number>;
    TreatmentBMPID?: FormControl<number>;
    FundingEventTypeID?: FormControl<number>;
    Year?: FormControl<number>;
    Description?: FormControl<string>;
    FundingEventFundingSources?: FormControl<Array<FundingEventFundingSourceSimpleDto>>;
    DisplayName?: FormControl<string>;
}

export class FundingEventDtoFormControls { 
    public static FundingEventID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FundingEventTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Year = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Description = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FundingEventFundingSources = (value: FormControlState<Array<FundingEventFundingSourceSimpleDto>> | Array<FundingEventFundingSourceSimpleDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<FundingEventFundingSourceSimpleDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static DisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}